Extract Device Stream Tests into a seperate unit test file
